Reworked Chat Queue
Made the chat queue its own object instead of trying to be fancy and do the chat log and chat queue as one object. Chat queue is now only the chats sent from this GUI / user while the chat log is the 512 most recent chats that this GUI / user is aware of.