Added start button and changed round resets
Start button appears on main canvas for the player who is drawing. This button sets playing to true, makes the button disappear, and allows the player to draw on the canvas. Round resets ( giveWord(boolean, String) ) sets playing to false, clears the canvas (and paintQueue), changes paint button visibility, and makes the start button visible if the player is the painter.