Made the main menu, fixed stuff in draw window
Simple menu now exists, when a game begins, it will create the drawing panel. Any info you want the drawing window to have should be fed to the menu panel, and it will hand it off to the drawing window when it is created. The drawing window is accessible from the menu window / thread.