Fixed Main Menu and GUI interactions
So, I made sure that the Main Menu GUI, the game GUI, and an outside program (the "controller") could interact properly. I separated out the creation of the game GUI from the Main Menu GUI so that it would be controlled by the "controller". The GUIRunner class has a suggested format for how the controller class could interact with the GUI. I've tested it to make sure it can do the things it should be able to do including: create a game; create a game, exit the game, and create a new game; create a game, start the game, exit the game, and create a new game; create a game, finish the round, begin a new round; all that stuff, but with join a game. I did not get to test multiple players, and so the ability to test between rounds for join game was limited since it needed the painter / host to press the start button.