Reduced errors to SocketClosed in MultiClient
These changes modify the code so that MultiClient only has SocketClosed for both the input and output classes. The errors on the server side are still broken pipe and EOF, which makes me think that it could be something with the way the server is trying to connect. I was going to play around with the server files, but server still references Client, and I am waiting on feedback.