addChat() now wraps; thanks recursion!
Finally remembered and implemented the wrapping function for the chat list. JList didn't like the "\n", and the only way to really do the multiple lines thing was with recursion. It will likely break words in inconvenient locations, but this version takes care of people who don't / won't use the space character.